+++ src/main/java/com/back/jsp/board/boundedContext/article/service/ArticleService.java
@@ -20,14 +20,14 @@ public class ArticleService {
     public Article writeArticle(String title, String content) {
         return repository.saveArticle(new Article(title, content));
     }
-    public Article getArticleById(int id) {
+    public Article getArticleById(long id) {
         Article article = repository.getArticleById(id);
         if (article == null) {
             return null;
         }
         return article;
     }
-    public Article getDetailById(int id) {
+    public Article getDetailById(long id) {
         Article article = getArticleById(id);
         if (article == null) {
             return null;
